requirements.txt
fastapi~=0.128.0
pydantic~=2.10.5
aiohttp~=3.12.14
yarl~=1.20.1
CairoSVG~=2.8.2
aiofiles~=24.1.0
emoji~=2.14.1
beautifulsoup4~=4.13.4
cryptography~=45.0.4
pillow~=11.3.0
orjson~=3.10.18
uvicorn~=0.34.3
aiosqlite~=0.21.0
qrcode~=8.2
validators~=0.35.0
questionary~=2.1.0